+++ CHANGELOG.md
@@ -1,16 +1,22 @@
 # LAST CHANGES:
 
-Version `v1.5.0 (8 Nov'23)`
+Version `v1.5.1 (15 Nov'23)`
 
+  - BUGFIX: env(TCLLIBPATH) cleared wrongly in apavebase.tcl
   - BUGFIX: Preferences/Projects: enable/disable translation options
   - NEW   : es.msg
+  - NEW   : Find/Replace at start if present at finish
   - NEW   : check for changes of Preferences' options
   - NEW   : file number in Projects/Files tab
+  - NEW   : !!! in Projects TODO makes its balloon "eternal" (hidden at clicks)
+  - CHANGE: TODO balloon not hidden at entering Projects
+  - CHANGE: run the loupe from the color picker
+  - CHANGE: preview.tcl: pickers
   - CHANGE: labels & buttons of Preferences & Projects
   - CHANGE: don't close "Projects/Files/No name"
   - CHANGE: Help/About/Packages tab
   - CHANGE: alited.ini
-  - CHANGE: packages: e_menu/apave 4.2.0, tkcc
+  - CHANGE: packages: e_menu/apave 4.2.0, tkcc, aloupe
 
 
 Version `v1.4.6 (25 Oct'23)`